+++ core/src/ons/util.js
@@ -352,4 +352,16 @@ util.isInteger = (value) => {
     Math.floor(value) === value;
 };
 
+/**
+ * @return {Obejct} Deferred promise.
+ */
+util.defer = () => {
+  const deferred = {};
+  deferred.promise = new Promise((resolve, reject) => {
+    deferred.resolve = resolve;
+    deferred.reject = reject;
+  });
+  return deferred;
+};
+
 export default util;