+++ CHANGELOG.md
@@ -4,13 +4,15 @@ CHANGELOG
 
 v2.0.2
 ----
- * switch: Fix regression where `change` event was not being triggered in some cases.
+ * ons-switch: Fix regression where `change` event was not being triggered in some cases.
  * angular1: Fix [#1609](https://github.com/OnsenUI/OnsenUI/issues/1609).
  * ons-toolbar: transparent modifier supports MD.
  * ons-toolbar-button: Style refactor.
  * ons-navigator: Fix show-init event order.
  * ons-tab: Add `badge` attribute to display notification on tab.
+ * ons.notification: Added `options.inputType` and `options.class`.
  * ons.notification: Fix [#1638](https://github.com/OnsenUI/OnsenUI/issues/1638).
+ * ons.notification: Fixed memory leak.
  * angular1: Fix [#1620](https://github.com/OnsenUI/OnsenUI/issues/1620).
  * ons-icon: Fix [#1636](https://github.com/OnsenUI/OnsenUI/issues/1636).